+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED](https://github.com/hahn-th/homematicip-rest-api/compare/2.7.0..master)
 
+### Added
+
+- Add support for HmIP-FWI (Wiegand Interface)
+
 ## [2.7.0](https://github.com/hahn-th/homematicip-rest-api/compare/2.6.0..2.7.0)
 
 ### Added